Fix authorization handling
`axios` changed how the `auth` and `headers` work. If `auth` was set then an `authorization` header would not be effective.

Refactored so that we set the `authorization` header manually based on either basic auth or digest auth flow.

Also refactored `requestConfig` to a class property that is not manipulated and passed around.